+++ src/pages/JHkim/Components/CommentLi.scss
@@ -1,0 +1,25 @@
+.new-li {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+
+  span strong {
+    font-weight: 900;
+  }
+
+  #delete-button {
+    border: none;
+    font-size: 16px;
+    font-weight: 900;
+    color: white;
+    background-color: white;
+  }
+
+  #delete-button:hover {
+    font-size: 17px;
+    color: red;
+    cursor: pointer;
+  }
+
+  /*댓글 li 좋아요 button style*/
+}
