+++ src/pages/JHkim/Components/Feed.scss
@@ -47,32 +47,6 @@
 .article-comment {
   padding: 10px;
 
-  .new-li {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-
-    span strong {
-      font-weight: 900;
-    }
-
-    #delete-button {
-      border: none;
-      font-size: 16px;
-      font-weight: 900;
-      color: white;
-      background-color: white;
-    }
-
-    #delete-button:hover {
-      font-size: 17px;
-      color: red;
-      cursor: pointer;
-    }
-
-    /*댓글 li 좋아요 button style*/
-  }
-
   input {
     width: 470px;
     outline: none;
